refactor(anima_train): 抽 dataset + collate + CLI 到 training/
ADR 0003 PR-A 第 3 步：

- training/dataset.py  ← BucketManager / ImageDataset / RepeatDataset /
                         MergedDataset / BucketBatchSampler /
                         CachedLatentDataset + collate_fn / collate_fn_cached
- training/cli.py      ← parse_args（被 test_anima_train_migration 直接调）+
                         _try_rich / _ask_* / _guess_default_paths / prompt_for_args

ImageDataset 里 caption_utils 动态加载的 `Path(__file__).parent`、
_guess_default_paths 里的 `Path(__file__).resolve().parent` 都加一层 .parent
保持等价语义（原指向 runtime/，搬到 runtime/training/ 后需多上一层）。注：
caption_utils 路径在原代码就指向不存在的 runtime/utils/，恒走 TXT 兜底；不在
本 PR 范畴内修。

anima_train.py 行数 1731 → 1042；migration test 全绿。
